Mobs now move around others fairly well
Although they cut corners now and when adding collision to the tiles they get stuck. Need to find a workaround for "soft" edges on the map to give the mobs more room to move while keeping them in the lane for the most part. Currently the Layer/Mask are turned off for the tilemap and staticbody on the tileset